requirements.txt
aioredis==1.3.1
async-timeout==3.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptg==0.2.post2
hiredis==1.1.0
idna==2.10
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
Telethon==1.20
urllib3==1.26.3
